fix loading screen movement bug
When you clicked on a next or prev day, and then moved the mouse around a lot, the panes would slide the viewport back over the current day, this happened because the next and prev day links were still present but opacity: 0, and the clear_map_classes function was attached to their mouseleave event. So when you were loading a new day, and you moved your mouse over the hidden links, the event would remove all the load-yesterday or load-tomorrow classes from the maps, and incorrectly move the viewport. This commit removes the loading map classes from the positioning map classes, and calls them each from the appropriate place.