correctly convert timestamps to UTC
whatever time you store in the timestamp is assumed to be UTC e.g.
4:28pm is assumed to be 4:28pm UTC even if it's EST. So we need to
convert the current timezone timestamp to UTC

this isn't a complete solution. we'd need to get the user's iOS timezone
to actually know what timezone their texts are in... this is close
enough for now though